vcで動きません

+++ clammbon/include/sort_algorithm/compare.hpp
@@ -8,6 +8,6 @@ point_type ul_choose(compared_type const& comp, point_type const ur, point_type 
 point_type dr_choose(compared_type const& comp, point_type const ul, point_type const ur, point_type const dl);
 point_type dl_choose(compared_type const& comp, point_type const ul, point_type const ur, point_type const dr);
 uint_fast64_t form_evaluate(compared_type const& comp_,std::vector<std::vector<point_type> > const& matrix);
-int get_kind_num(question_raw_data const& data_,std::vector<std::vector<point_type>> const& matrix, int const x, int const y);
+size_t get_kind_num(question_raw_data const& data_, std::vector<std::vector<point_type>> const& matrix, int const x, int const y);
 
 #endif
